migrations

+++ sql/migrations/018_create_jobs.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS jobs (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     job_key VARCHAR(64) NOT NULL UNIQUE,
     title_ru VARCHAR(120) NOT NULL,
     title_en VARCHAR(120) NULL,
@@ -11,11 +11,11 @@ CREATE TABLE IF NOT EXISTS jobs (
     reward_amount INT NOT NULL DEFAULT 0,
     cooldown_seconds INT NOT NULL DEFAULT 0,
     enabled BOOLEAN NOT NULL DEFAULT TRUE,
-    reward_item_id BIGINT NULL,
+    reward_item_id INT NULL,
     reward_item_chance_percent DECIMAL(5, 2) NULL,
     reward_item_quantity INT NOT NULL DEFAULT 1,
-    created_by_member_id BIGINT NULL,
-    updated_by_member_id BIGINT NULL,
+    created_by_member_id INT NULL,
+    updated_by_member_id INT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     CONSTRAINT fk_jobs_reward_item
@@ -33,8 +33,8 @@ CREATE TABLE IF NOT EXISTS jobs (
 CREATE INDEX idx_jobs_enabled_updated_at ON jobs(enabled, updated_at DESC);
 
 CREATE TABLE IF NOT EXISTS member_job_cooldowns (
-    member_id BIGINT NOT NULL,
-    job_id BIGINT NOT NULL,
+    member_id INT NOT NULL,
+    job_id INT NOT NULL,
     last_run_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (member_id, job_id),
     CONSTRAINT fk_member_job_cooldowns_member
